requirements.txt
api==0.0.7
asgiref==3.2.10
astroid==2.4.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nose==1.3.7
PyJWT==1.7.1
pylint==2.5.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
